Add blender for boolean operations
This commit adds the possibility to use Blender instead of OpenSCAD
for boolean operations on meshes. Blender is way faster than OpenSCAD
when it comes to big meshes like the Lithophane workbench produces.

The processor exports all meshes to stl into a temporary folder. Than it
writes a json file that defines the operations to execute. Afterwards it
calls blender with a python script that reads the config and the meshes
and applies a boolean modifier to them. The script then exports the final
mesh to a file in the temporary folder. At the end the processor reads
the final mesh and deletes the temporary folder.

The blender processor can be enabled or disabled via preferences by the
user.

optimizes #33